move ethclient to plugin/evm
