+++ RELEASES.md
@@ -5,7 +5,7 @@
 - Major refactor to use [`libevm`](https://github.com/ava-labs/libevm) for EVM execution, database access, types & chain configuration. This improves maintainability and enables keeping up with upstream changes more easily.
 - Add metrics for ACP-176
 - Removed the `"price-options-max-base-fee"` config flag
-- Moved extra types returned in `ethclient` package to a new package `ethclient/customclient` which supports same functionality as `ethclient` but with the new types registered in header and block.
+- Moved extra types returned in `ethclient` package to a new package `plugin/evm/customethclient` which supports same functionality as `ethclient` but with the new types registered in header and block.
 
 ## [v0.15.0](https://github.com/ava-labs/coreth/releases/tag/v0.15.0)
