+++ plugin/evm/customethclient/ethclient_ext.go
@@ -1,7 +1,7 @@
 // (c) 2019-2025, Ava Labs, Inc. All rights reserved.
 // See the file LICENSE for licensing terms.
 
-package customclient
+package customethclient
 
 import (
 	"context"
