+++ plugin/evm/customethclient/example_test.go
@@ -1,4 +1,4 @@
-package customclient
+package customethclient
 
 import (
 	"context"
